pyproject.toml
[project]
name = "edgar"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "datetime>=5.5",
    "duckdb>=1.2.2",
    "fredapi>=0.5.2",
    "ipykernel>=6.29.5",
    "logging>=0.4.9.6",
    "nbformat>=5.10.4",
    "pandas>=2.2.3",
    "pandas-datareader>=0.10.0",
    "pathlib>=1.0.1",
    "plotly>=6.0.1",
    "pyarrow>=19.0.1",
    "pytest>=9.0.1",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "sec-cik-mapper>=2.1.0",
    "tabulate>=0.9.0",
    "torch>=2.9.1",
    "tqdm>=4.67.1",
    "transformers>=4.57.1",
    "typing>=3.10.0.0",
    "yfinance>=0.2.55",
]
